+++ ModHelperData.cs
@@ -2,16 +2,16 @@ namespace TacticalTweaks;
 
 public static class ModHelperData
 {
-    public const string WorksOnVersion = "55.3";
-    public const string Version = "1.2.2";
+    public const string WorksOnVersion = "55.2";
+    public const string Version = "1.2.3";
     public const string Name = "Tactical Tweaks";
 
     public const string Description =
         """
         Adds an assortment of minor gameplay altering tweaks to BTD6, including **Buffable Paragons**!. See the Main Readme for full details.
 
-        New: Beast Handler Lone T5s, Brickell Mermonkeys, Elite Defender Target Switch, Etienne Targeting, Sticky Bomb Attack Speed, Better Laser Blasts, Better Trojan, Better Monkey City, Better Central Market, Auto Collect/Deposit Banks, Obyn Price Tweak, More Overclock Interactions, More Unfreezable Towers
-        Formerly Useful Utilities: Dartling/Mortary/Sentry Targeting, Auto Trap Refresh
+        New: Beast Handler Lone T5s, Brickell Mermonkeys, Elite Defender Target Switch, Etienne Targeting, Sticky Bomb Attack Speed, Better Laser Blasts, Better Trojan, Better Monkey City, Better Central Market, Auto Collect/Deposit Banks, Obyn Price Tweak, More Overclock Interactions, More Unfreezable Towers, More Pyrotechnics Towers
+        Formerly Useful Utilities: Dartling/Mortar/Sentry Targeting, Auto Trap Refresh
         Old Mods: Better Armor Piercing, Better Ezili Totem
         """;